\name{apo.data.vm}
\alias{apo.data.vm}

\docType{data}
\title{Apo Data set}
\description{
An exemple data set for unpaired data analysis.
A list with the normalized Apo data set as used in the original article.
}
\usage{data(apo.data.vm)}
\format{
apo.data.vm is a list with 3 elements :
\item{apo.geneid}{A vector with gene names}
\item{apo.cond1}{A Matrix with 6226 rows and 8 columns with normalized normal mice measurements}
\item{apo.cond2}{A Matrix with 6226 rows and 8 columns with normalized KO mice measurements}
}
\details{

}
\source{
}
\references{
M.J. Callow, S. Dudoit, E.L. Gong, T.P. Speed, and E.M. Rubin. Microarray expression profiling 
identifies genes with altered expression in hdl-deficien mice. Genome Res., 10(12) : 2022-9, 2000
}
\examples{
data(apo.data.vm)
}
\keyword{datasets}
