\name{badqual.mat}
\alias{badqual.mat}

\title{How to get bad quality matrix}
\description{ This function extracts the matrix
with quality measures from a gene expression data object
}
\usage{
badqual.mat(data)
}

\arguments{
  \item{data}{Gene expression data object}
}
\details{

}
\value{
}
\references{}
\author{Paul Delmar}
\note{} 
\seealso{}
\examples{
}
\keyword{internal}
