\name{block.vect}
\alias{block.vect}

\title{Extraction of the block information vector}
\description{ This function extracts a vector with block (print-tip)
information from the gene expression data object
 
}
\usage{
block.vect(data)
}

\arguments{
  \item{data}{Gene expression data object}
}
\details{

}
\value{
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
