\name{boxplotvm.g}
\alias{boxplotvm.g}

\title{Boxplot of mean log-intensity in each variance group }
\description{
 Boxplot of mean log-intensity in each variance group 
}
\usage{
boxplotvm.g(x)
}

\arguments{
  \item{x}{Gene expression data object}
}
\details{
The y axis is the standard deviation of the gene.
}
\value{A boxplot of the distribution of mean log-intensity in each variance group}
\references{}
\author{Paul Delmar}
\note{}

\seealso{\code{\link{vm.analysis}},\code{\link{vm.analysis.paired}},
        \code{\link{plotrm}},\code{\link{qqplot.vm}},\code{\link{plotvm}}}
 }
\examples{}
\keyword{htest}
