\name{comp.bic}
\alias{comp.bic}

\title{BIC and AIC criteria}
\description{ This function computes the value of AIC and BIC criteria
for 2 to 10 variance components
}
\usage{
comp.bic(VAR,df,stop.crit=1.e-6,n.max=20,display=FALSE)
}

\arguments{
  \item{VAR}{Vector of estimated variance}
  \item{df}{degrees of freedom of the estimated variance}
  \item{stop.crit}{ Real : convergence criterion for the EM algo}
  \item{n.max}{Integer : maximum number of iterations}
  \item{display}{Boolean : Should the program write its results on the console}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
