\name{comp.bic.stoch}
\alias{comp.bic.stoch}

\title{Computes mixture parameters for varying number of components }
\description{
Computes mixture parameters for varying number of components
using random initialisation of EM algorithm
}
\usage{
comp.bic.stoch(VAR, df, stop.crit = 1e-06, n.max = 20)

}

\arguments{
  \item{VAR}{Vector of estimated variance}
  \item{df}{degrees of freedom of the estimated variance}
  \item{stop.crit}{ Real : convergence criterion for the EM algo}
  \item{n.max}{Maximum number of mixtures to consider}
}
\details{

}
\value{
}
\references{}
\author{Paul Delmar}
\note{}
\examples{}
\keyword{internal}
