\name{cond1.mat}
\alias{cond1.mat}

\title{Extraction of the matrix with normalized expression data in condition 1}
\description{This function extracts the matrix with normalized expression data in condition 1
 
}
\usage{
cond1.mat(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
