\name{cond2.mat}
\alias{cond2.mat}

\title{Extraction of the matrix with normalized expression data in condition 2}
\description{Extracts the matrix with normalized expression data in condition 2
 
}
\usage{
cond2.mat(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{




## The function is currently defined as
function(data)
{
    data$cond2
  }
}
\keyword{internal}
