\name{df.vect}
\alias{df.vect}
\title{Number of degrees of freedom used in the variance estimation}
\description{
 Returns vector of the number of degrees of freedom used in the variance estimation
}
\usage{
df.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{
A vector : the number of degrees of freedom for the estimation of the variance
for each gene}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{}
\keyword{internal}
