\name{dif.f}
\alias{dif.f}

\title{Intermediary function}
\description{This function is involved in the computation of the variance
of the estimated parameters (variance mixture model) 
}
\usage{
dif.f(x, nmixt, pi, b, c)
}

\arguments{
  \item{x}{  }
  \item{nmixt}{ number of groups of variance  }
  \item{pi}{  }
  \item{b}{ parameter of the gamma function }
  \item{c}{ parameter of the gamma function }
}
\details{
}
\value{ 
}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{
}
\keyword{internal}
