\name{fdr.nmatch.an.anova}
\alias{fdr.nmatch.an.anova}
\title{Match result for homoscedastic model}
\description{}
\usage{
fdr.nmatch.an.anova(dat1,dat2,qval=0.05,lambda=seq(0,0.95,0.05))
}
\arguments{
  \item{dat1}{}
  \item{dat2}{}
  \item{qval}{}
  \item{lambda}{}
}
\details{
}
\value{}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
