\name{find.gene.index}
\alias{find.gene.index}
\title{Indexes of a vector of gene IDs}
\description{This function finds indexes of a vector of gene IDs
}
\usage{
find.gene.index(data, gene)
}
\arguments{
  \item{data}{gene expression data object}
  \item{gene}{Character vector with gene ids}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}


\seealso{}
\examples{}
\keyword{internal}
