\name{gene.anot.data.frame}
\alias{gene.anot.data.frame}
\title{Gene anotation data frame}
\description{
This function extracts the gene anotation data.frame
}
\usage{
gene.anot.data.frame(data)
}
\arguments{
  \item{data}{gene expression data object}
}
\details{
}
\value{
A data.frame of gene anotations
}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{
}
\keyword{internal}
