\name{geneid.vect}
\alias{geneid.vect}

\title{Extraction of a vector with the name of each gene}
\description{This function extracts a vector with the name of each gene
 
}
\usage{
geneid.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
