\name{get.sas.export}
\alias{get.sas.export}

\title{get.sas.export}
\description{
 
}
\usage{
get.sas.export(file.test, file.cont, file.qual)
}

\arguments{
  \item{file.test}{ file containing test condition }
  \item{file.cont}{ file containing control condition }
  \item{file.qual}{  }
}
\details{

}
\value{
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
