\name{group.vect}
\alias{group.vect}

\title{Extraction of the vector of the variance group of each gene}
\description{This function extracts the vector of the variance group of each gene.
A variance group is determined by the variance mixture model.
 
}
\usage{
group.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{




## The function is currently defined as
function(data)
{
  data$stat2$group
  }
}
\keyword{internal}
