\name{histresid.vm}
\alias{histresid.vm}

\title{histresid.vm}
\description{
 
}
\usage{
histresid.vm(x,...)
}

\arguments{
  \item{x}{gene expression data object}
  \item{\dots}{ }
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
