\name{histvm}
\alias{histvm}

\title{histvm}
\description{
 }
\usage{
histvm(x, ...)
}

\arguments{
  \item{x}{gene expression data object}
  \item{\dots}{further arguments to the hist function}
}
\details{
}
\value{

}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{
}
\keyword{internal}
