\name{keep.index}
\alias{keep.index}

\title{Subset of vm.result data object based on index}
\description{This function extracts a subset of vm.result data object based on index.

}
\usage{
keep.index(data, index)
}

\arguments{
  \item{data}{gene expression data object}
  \item{index}{vector of gene indexes}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}

 

\seealso{}
\examples{}
\keyword{internal}
