\name{logratio.mat}
\alias{logratio.mat}

\title{Matrix of normalized log-ratio}
\description{This function extracts the matrix of normalized log-ratio
 
}
\usage{
logratio.mat(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
