\name{logratio.vect}
\alias{logratio.vect}

\title{Vector of mean log-ratio}
\description{This function extracts the vector of mean log-ratio
 
}
\usage{
logratio.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
