\name{make.param}
\alias{make.param}

\title{A function to build a parameter object }
\description{A function to build a parameter object}
\usage{
make.param(VAR, df, sd.init, pi.init, nmixt)
}

\arguments{
  \item{VAR}{vector of estimated variance}
  \item{df}{degrees of freedom of the estimated variance}
  \item{sd.init}{vector of initial sd (mixture model)}
  \item{pi.init}{vector of initial pi (mixture model)}
  \item{nmixt}{number of components in the mixture}
}
\details{}
\value{}
\references{ }
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
