\name{meanint.mat}
\alias{meanint.mat}

\title{Matrix of gene mean intensity}
\description{This function extracts the matrix of gene mean intensity
 
}
\usage{
meanint.mat(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
