\name{meanint.vect}
\alias{meanint.vect}

\title{Vector of mean gene intensity}
\description{This function extracts the vector of mean gene intensity
}
\usage{
meanint.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{
}
\keyword{internal}
