\name{n.genes}
\alias{n.genes}

\title{Number of genes in a gene expression data object}
\description{Number of genes in a gene expression data object
 
}
\usage{
n.genes(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{This function returns the number of analyzed gene probes
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{\code{\link{vm.analysis}},\code{\link{vm.analysis.paired}}}
\examples{
}
\keyword{htest}
