\name{n.mixt}
\alias{n.mixt}
\title{Number of mixture components in the model}
\description{
}
\usage{
n.mixt(data)
}
\arguments{
  \item{data}{gene expression data object }
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{}
\keyword{internal}
