\name{nmatch.an.anova}
\alias{nmatch.an.anova}
\title{ Match result for homoscedastic model}
\description{}
\usage{
nmatch.an.anova(dat1, dat2, P = 0.05)
}
\arguments{
  \item{dat1}{}
  \item{dat2}{}
  \item{P}{}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
