\name{nmatch.an.gene}
\alias{nmatch.an.gene}
\title{Match result for gene-specific model }
\description{

}
\usage{
nmatch.an.gene(dat1, dat2, P = 0.05)
}
\arguments{
  \item{dat1}{}
  \item{dat2}{}
  \item{P}{}
}
\details{
}
\value{}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{}
\keyword{internal}
