\name{nmatch.an}
\alias{nmatch.an}
\title{Match results for Varmixt model}
\description{}
\usage{
nmatch.an(dat1, dat2, P = 0.05)
}
\arguments{
  \item{dat1}{ }
  \item{dat2}{}
  \item{P}{}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}


\seealso{}
\examples{}
\keyword{internal}
