\name{parameters.data.frame}
\alias{parameters.data.frame}

\title{Data.frame with the variance mixture model parameters}
\description{This function extracts a data.frame with the variance mixture model parameters
}
\usage{
parameters.data.frame(data)
}

\arguments{
  \item{data}{Gene expression data object}
}
\details{
}
\value{
A data frame with the estimated value of the mixture model parameters. 
}
\references{}
\author{Paul Delmar}
\note{}

\seealso{\code{\link{vm.analysis}},\code{\link{vm.analysis.paired}}}
\examples{
}
\keyword{internal}
