\name{plotsdm}
\alias{plotsdm}

\title{plotsdm}
\description{
 
}
\usage{
plotsdm(x,colors=TRUE)
}

\arguments{
  \item{x}{gene expression data object}
  \item{colors}{Logical. If TRUE a different color is used for each variance group.}
}
\details{}
\value{A plot of log of mean log-ratio standard deviation versus mean log-intensity}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
