\name{plotsdt}
\alias{plotsdt}

\title{ Plot Test statistic versus denominator of test statistic}
\description{}
\usage{
plotsdt(x,pval.f,test.stat.f,pval=0.05)
}

\arguments{
  \item{x}{gene expression data object}
  \item{pval.f}{ }
  \item{test.stat.f}{ }
  \item{pval}{pvalue cut-off}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{htest}
