\name{plotsdt.VM}
\alias{plotsdt.VM}

\title{Plot of VM test statistic versus denominator}
\description{Plot VM test statistic versus denominator}
\usage{
plotsdt.VM(x,pval=0.05)
}

\arguments{
  \item{x}{gene expression data object}
  \item{pval}{p value cut-off for coloring points}
}
\details{}
\value{A plot of the denominator of the VM test statistic versus the test static}
\references{ }
\author{Paul Delmar}
\note{}

\seealso{}
\examples{}
\keyword{htest}
