\name{plotvm}
\alias{plotvm}
\title{Plot of log-variance versus mean intensity}
\description{
Plot of log-variance versus mean intensity
}
\usage{
plotvm(x,colors=TRUE)
}

\arguments{
  \item{x}{the results of vm.analysis of vm.analysis.paired function}
  \item{colors}{if TRUE a different color is used for each variance group}
 
}
\details{
 }
\value{
 Plot of log-variance versus mean intensity
}
\references{}
\author{Paul Delmar}
\note{}

\seealso{\code{\link{vm.analysis}},\code{\link{vm.analysis.paired}},
        \code{\link{plotrm}},\code{\link{qqplot.vm}},\code{\link{boxplotvm}}}
 }
\examples{
}
\keyword{htest}
