\name{prt.varmixt.res}
\alias{prt.varmixt.res}
\title{Parameters of varmixt analysis}
\description{
This function prints parameters of varmixt analysis}
\usage{
prt.varmixt.res(data)
}
\arguments{
  \item{data}{Gene expression data object. Result of varmixt analysis}
}
\details{

}
\value{
}
\references{ }
\author{Paul Delmar}
\note{}

\seealso{}
\examples{}
\keyword{internal}
