\name{pval.anova.corrected.vect}
\alias{pval.anova.corrected.vect}

\title{Vector of corrected (Bonferroni) p-value from homoscedastic variance model}
\description{This function extracts the vector of corrected (Bonferroni) p-value from homoscedastic variance model}
\usage{
pval.anova.corrected.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{The vector of corrected (Bonferroni) p-value from homoscedastic variance model}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
