\name{pval.anova.vect}
\alias{pval.anova.vect}

\title{Vector of p-value from homoscedastic variance model}
\description{This function extracts the vector of p-value from homoscedastic variance model
 
}
\usage{
pval.anova.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
