\name{pval.gene.corrected.vect}
\alias{pval.gene.corrected.vect}

\title{Vector of corrected (Sidak) p-value from gene-specific variance model}
\description{this function extracts the vector of corrected (Sidak) p-value from gene-specific variance model
 
}
\usage{
pval.gene.corrected.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
