\name{pval.VM.corrected.vect}
\alias{pval.VM.corrected.vect}
\title{Bonferroni corrected p-value from "continuous" mixture model (without assignment)}
\description{
This function extracts the Bonferroni corrected p-value from "continuous" mixture model (without assignment)}
\usage{
pval.VM.corrected.vect(data)
}
\arguments{
  \item{data}{gene expression data object}
}
\details{
}
\value{
The  vector of the Bonferoni corrected p-value from "continuous" mixture model (without assignment)
}
\references{}
\author{}
\note{}
\seealso{\code{\link{fdr.an}},\code{\link{pval.an}}}
\examples{}
\keyword{internal}
