\name{pval.VM2.corrected.vect}
\alias{pval.VM2.corrected.vect}

\title{Vector of corrected (Sidak) p-value from variance mixture model}
\description{This function extracts the vector of corrected (Sidak) p-value from variance mixture model
 
}
\usage{
pval.VM2.corrected.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
