\name{pval.VM2.vect}
\alias{pval.VM2.vect}

\title{Vector of p-value from variance mixture model}
\description{This function extracts the vector of p-value from variance mixture model
 
}
\usage{
pval.VM2.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{} 
\seealso{}
\examples{
}
\keyword{internal}
