\name{qplot.vm}
\alias{qplot.vm}
\title{Plot Diagnostic of FDR computation}
\description{Plot Diagnostic of FDR computation
}
\usage{
qplot.vm(data, lambda = seq(0, 0.95, 0.05))
}
\arguments{
  \item{data}{gene expression data object}
  \item{lambda}{lambda parameter of qvalue function}
}
\details{}
\value{Plot Diagnostic of FDR computation}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{}
\keyword{htest}
