\name{qqplot.gamma}
\alias{qqplot.gamma}

\title{QQ plot of gamma distribution}
\description{
QQ plot of gamma distribution
}
\usage{
qqplot.gamma(vars, df, PLOT = TRUE, ...)
}
\arguments{
  \item{vars}{vector of estimated variance}
  \item{df}{degrees of freedom of the estimated variance}
  \item{PLOT}{boolean}
  \item{\dots}{supplemental argument to the plot function}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
