\name{qqplot.var.vm.2}
\alias{qqplot.var.vm.2}

\title{QQ plot of the variance versus gamma distribution in one variance group}
\description{
QQ plot of the variance versus gamma distribution in one variance group
}
\usage{
qqplot.var.vm.2(data, i)
}
\arguments{
  \item{data}{gene expression data object}
  \item{i}{component number}
}
\details{}
\value{
QQ plot of the distribution of estimated variance of genes assigned to group i versus 
the theoretical quantiles of a gamma distribution.

}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{htest }
