\name{qqplot.var.vm}
\alias{qqplot.var.vm}
\title{QQ plot of the distribution of estimated variance}
\description{QQ plot of the distribution of estimated variance}
\usage{
qqplot.var.vm(data)
}
\arguments{
  \item{data}{Gene expression data object}
}
\details{
 
}
\value{
Group by group qq-plot of the distribution of estimated variance versus a theoretical 
Gamma distribution
}
\references{}
\author{Paul Delmar}
\note{}



\seealso{ }
\examples{}
\keyword{htest}
