\name{qqplot.vm.2}
\alias{qqplot.vm.2}

\title{qqplot.vm.2}
\description{
 
}
\usage{
qqplot.vm.2(data,i,...)
}

\arguments{
  \item{data}{gene expression data object}
  \item{i}{ }
  \item{\dots}{optional parameters }
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
