\name{qual.vect}
\alias{qual.vect}

\title{Vector with global quality info for each gene}
\description{This function extracts the vector with global quality info for each gene
 
}
\usage{
qual.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
