\name{quality.filter}
\alias{quality.filter}

\title{Filter out genes with low quality observations}
\description{Filter out genes with low quality observations
 
}
\usage{
quality.filter(data, badtol)
}

\arguments{
  \item{data}{gene expression data object}
  \item{badtol}{  }
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
