\name{rank.anova.vect}
\alias{rank.anova.vect}

\title{Vector of the ranks of each gene with homoscedastic variance model}
\description{This function extracts the vector of the ranks of each gene with homoscedastic variance model
 
}
\usage{
rank.anova.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
