\name{rank.gene.vect}
\alias{rank.gene.vect}

\title{Vector of the ranks of each gene with gene-specific variance model}
\description{This function extracts the vector of the ranks of each gene with gene-specific variance model
 
}
\usage{
rank.gene.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
