\name{rank.VM2.vect}
\alias{rank.VM2.vect}

\title{Vector of the ranks of each gene with variance mixture model}
\description{This function extracts the vector of the ranks of each gene with variance mixture model
 
}
\usage{
rank.VM2.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
