\name{raw.cond1.mat}
\alias{raw.cond1.mat}

\title{Matrix of raw data in condition 1}
\description{This function extracts the matrix of raw data in condition 1
 
}
\usage{
raw.cond1.mat(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
