\name{raw.log.ratio.mat}
\alias{raw.log.ratio.mat}
\title{Matrix of raw log-ratio}
\description{
This function extracts the matrix of raw log-ratio
}
\usage{
raw.log.ratio.mat(data)
}

\arguments{
  \item{data}{Gene expression data object}
}
\details{
}
\value{
A matrix with raw log-ration (one gene per row, one replicate per column)
}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{
}
\keyword{internal}
