\name{raw.meanint.mat}
\alias{raw.meanint.mat}
\title{Matrix of raw mean log-intensity}
\description{
 This function extracts the matrix of raw mean log-intensity
}
\usage{
raw.meanint.mat(data)
}

\arguments{
  \item{data}{A gene expression data object}
}
\details{
}
\value{
A matrix with raw mean log-intensity (one gene per row, one replicate per column)
}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{
}
\keyword{internal}
