\name{remove.index}
\alias{remove.index}

\title{Subset of vm.result object by removing index}
\description{This function extracts a subset of vm.result object by removing index}
\usage{
remove.index(data, index)
}

\arguments{
  \item{data}{gene expression data object}
  \item{index}{vector of integer : Indexes of genes to remove }
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}


\seealso{}
\examples{}
\keyword{internal}
