\name{res.data.frame}
\alias{res.data.frame}

\title{Dataframe with analysis results}
\description{This function builds a dataframe with analysis results
 
}
\usage{
res.data.frame(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
