\name{residual.mat}
\alias{residual.mat}

\title{Matrix of residuals}
\description{This function extracts the matrix of residuals
 
}
\usage{
residual.mat(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
