\name{sbset.gene}
\alias{sbset.gene}

\title{Subset of vm.result object using gene IDs}
\description{This function extracts the subset of vm.result object using gene IDs}
\usage{
sbset.gene(data, gene)
}

\arguments{
  \item{data}{gene expression data object}
  \item{gene}{a vector of gene names matchin the gene id of the data object}
}
\details{}
\value{A subset of the data with only those genes also in the gene argument.}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{htest}
