\name{sd.delta.gene.vect}
\alias{sd.delta.gene.vect}

\title{Vector of standard deviations of the the test statistic in gene-specific model}
\description{This function extracts the vector with standard deviation of the mean difference in gene-specific model

}
\usage{
sd.delta.gene.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
