\name{sd.delta.VM.vect}
\alias{sd.delta.VM.vect}

\title{Standard deviation of the test statistic in the variance mixture model}
\description{This function extracts the standard deviation of the test statistic in the variance mixture model
}
\usage{
sd.delta.VM.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{

  

 
 
}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
