\name{sd.delta.VM2.vect}
\alias{sd.delta.VM2.vect}

\title{Standard deviation of the test statistic in the variance mixture model}
\description{This function extracts the standard deviation of the test statistic in the variance mixture model
}
\usage{
sd.delta.VM2.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\references{}
\author{Paul Delmar}

\keyword{internal}
