\name{sd.g.VM2.vect}
\alias{sd.g.VM2.vect}
\title{Gene variance in variance mixture model}
\description{
This function extracts a vector of gene variance in variance mixture model
}
\usage{
sd.g.VM2.vect(data)
}
\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{Vector of gene variance in variance mixture model
}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
