\name{sd.param}
\alias{sd.param}

\title{Standard deviation of the variance mixture model parameters}
\description{This function computes the standard deviation of the variance mixture model parameters
}
\usage{
sd.param(data)
}

\arguments{
  \item{data}{Gene expression data object}
}
\details{
}
\value{
A data.frame with the value of the standard deviation of the mixture model parameters 
}
\references{}
\author{Paul Delmar}
\note{}

\seealso{}
\examples{
}
\keyword{htest}
