\name{sigma.delta.gene.vect}
\alias{sigma.delta.gene.vect}

\title{Vector of standard deviations of the the test statistic in gene-specific model}
\description{This function extracts the vector with standard deviation of the mean difference in gene-specific model
 
}
\usage{
sigma.delta.gene.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\author{Paul Delmar}
\keyword{internal}
