\name{sigmaV2.vect}
\alias{sigmaV2.vect}

\title{Vector the gene variance estimated with variance mixture model}
\description{This function extracts the vector the variance estimated with variance mixture model
 
}
\usage{
sigmaV2.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}

\author{Paul Delmar}
\keyword{internal}
