\name{stat1.data.frame}
\alias{stat1.data.frame}

\title{Dataframe with simple statistics}
\description{This function extracts the data.frame with simple statistics
}
\usage{
stat1.data.frame(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\author{Paul Delmar}
\note{}

 
\seealso{}
\keyword{internal}
