\name{stat2.data.frame}
\alias{stat2.data.frame}

\title{Data.frame with variance mixture model results}
\description{This function extracts the data.frame with variance mixture model results
 in a gene expression data object.
}
\usage{
stat2.data.frame(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\author{Paul Delmar}
\seealso{}
\examples{
}
\keyword{internal}
