\name{test.stat.anova.vect}
\alias{test.stat.anova.vect}

\title{Vector of test statistic from the homoscedastic variance model}
\description{Extracts the vector of test statistic from the homoscedastic variance model
in a gene expression data object.
}
\usage{
test.stat.anova.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{

}
\value{}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
