\name{test.stat.gene.vect}
\alias{test.stat.gene.vect}
\title{Vector of test statistic from the gene-specific variance model}
\description{This function extracts the vector of test statistic from the gene-specific variance model
 in a gene expression data object}
\usage{test.stat.gene.vect(data)}
\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{the vector of test statistic from the gene-specific variance model}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
