\name{test.stat.VM.vect}
\alias{test.stat.VM.vect}
\title{Continuous Variance Mixture Model test statistic}
\description{
This function extracts the continuous Variance Mixture Model test statistic}
\usage{
test.stat.VM.vect(data)
}
\arguments{
  \item{data}{gene expression data object}
}
\details{
}
\value{
A vector of the continuous Variance Mixture Model test statistic
}
\author{Paul Delmar}
\examples{}
\keyword{internal}
