\name{test.stat.VM2.vect}
\alias{test.stat.VM2.vect}

\title{Vector of test statisitic from the variance mixture model}
\description{This function
extracts the vector of test statisitic from the variance mixture model
in a gene expression data object}
\usage{
test.stat.VM2.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{The vector of test statisitic from the variance mixture model}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
