\name{var.estim.mat}
\alias{var.estim.mat}
\title{Covariance matrix of parameters estimates}
\description{
This function extracts the covariance matrix of parameters estimates
}
\usage{
var.estim.mat(data)
}

\arguments{
  \item{data}{A gene expression data object}
}
\details{}
\value{ Covariance matrix of parameters estimates}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
