\name{var.gene.vect}
\alias{var.gene.vect}

\title{Gene variance vector}
\description{This function extracts gene variance vector from the gene expression data object}
\usage{
var.gene.vect(data)
}

\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{A vector with the gene-specific variance of each gene
}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
