\name{var.to.vardelta.vect}
\alias{var.to.vardelta.vect}
\title{Gene log-intensity variance to mean difference variance}
\description{
 Gene log-intensity variance to mean difference variance
}
\usage{
var.to.vardelta.vect(data)
}
\arguments{
  \item{data}{gene expression data object}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}
\seealso{}
\examples{}
\keyword{internal}
