\name{varmat}
\alias{varmat}

\title{Variance matrix of the parameters estimates}
\description{This function computes the variance matrix
of the parameters estimated in the variance mixture model.
}
\usage{
varmat(vect, nmixt, pi, b, c)
}

\arguments{
  \item{vect}{  }
  \item{nmixt}{ }
  \item{pi}{ }
  \item{b}{  }
  \item{c}{}
}
\details{}
\value{}
\references{}
\author{Paul Delmar}
\note{}

 
\seealso{}
\examples{
}
\keyword{internal}
